requirements.txt
absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
deepface==0.0.75
Django==4.1.1
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
flatbuffers==2.0.7
gast==0.4.0
gdown==4.5.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.0
h5py==3.7.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
mtcnn==0.1.1
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
protobuf==3.19.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
retina-face==0.0.12
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
tqdm==4.64.1
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
